SkillRating.jl

SkillRating.elo_expected

SkillRating.elo_rating

SkillRating.glicko_RD

SkillRating.glicko_RD_increase

SkillRating.glicko_c

SkillRating.glicko_expected

SkillRating.glicko_rating

SkillRating.elo_expected — Method

elo_expected(RA, RB)

Arguments:

RA: Rating of player A

RB: Rating of player B

Calculate the expected outcome using elo. See: https://en.wikipedia.org/wiki/Eloratingsystem#Mathematical_details

SkillRating.elo_rating — Method

elo_rating(r, rjs::Vector, sjs::Vector; k = 32)

Arguments:

r: Rating of player

rjs: Ratings of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate the elo rating change. See: https://en.wikipedia.org/wiki/Eloratingsystem#Mathematical_details

SkillRating.glicko_RD — Method

glicko_RD(r, RD, rjs::Vector, RDjs::Vector, sjs::Vector)
glicko_RD(r, RD, rj::T, RDj::T, sj::T) where T <: Real

Arguments:

r: Rating of player

RD: Rating deviation of player

rjs: Ratings of oppenents within rating period.

RDjs: Rating devations of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate a new glicko rating deviation for games within a rating period or for a single game. See: http://www.glicko.net/glicko/glicko.pdf

SkillRating.glicko_RD_increase — Method

glicko_RD_increase(rd; c = 63.2, maxrd=350)

Arguments:

RD: Rating deviation at end of rating period.

c: Increase in uncertainty between rating periods.

maxRD: Rating deviation of a player who does not play.

Calculate rating deviation after the end of a rating period. See: http://www.glicko.net/glicko/glicko.pdf

SkillRating.glicko_c — Method

glicko_c(rating_perods_to_max, maxrd, medianrd)

Calculate c, the increase in uncertainty between rating periods. See: http://www.glicko.net/glicko/glicko.pdf

SkillRating.glicko_expected — Method

elo_expected(RA, RB)

Arguments:

RA: Rating of player A

RB: Rating of player B

Calculate the expected outcome using glicko. See: http://www.glicko.net/glicko/glicko.pdf

SkillRating.glicko_rating — Method

glicko_rating(r, RD, rjs::Vector, RDjs::Vector, sjs::Vector)
glicko_rating(r, RD, rj::T, RDj::T, sj::T) where T <: Real

Arguments:

r: Rating of player

RD: Rating deviation of player

rjs: Ratings of oppenents within rating period.

RDjs: Rating devations of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate a new glicko rating for games within a rating period or for a single game. See: http://www.glicko.net/glicko/glicko.pdf