SkillRating.jl

SkillRating.elo_expected

SkillRating.elo_rating

SkillRating.glicko1_to_glicko2

SkillRating.glicko2

SkillRating.glicko_RD

SkillRating.glicko_RD_increase

SkillRating.glicko_c

SkillRating.glicko_expected

SkillRating.glicko_rating

SkillRating.elo_expected — Method

elo_expected(RA, RB)

Arguments:

RA: Rating of player A

RB: Rating of player B

Calculate the expected outcome using elo. See: https://en.wikipedia.org/wiki/Eloratingsystem#Mathematical_details

SkillRating.elo_rating — Method

elo_rating(r, rjs::Vector, sjs::Vector; k = 32)

Arguments:

r: Rating of player

rjs: Ratings of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate the elo rating change. See: https://en.wikipedia.org/wiki/Eloratingsystem#Mathematical_details

SkillRating.glicko2 — Method

glicko2(μ::Real, ϕ::Real, σ::Real, μjs::Vector{Real}, ϕjs::Vector{Real}, sjs::Vector{Real}; τ=0.05, ϵ=10^-6)

Arguments:

μ: Rating of player

ϕ: Rating deviation of player

σ: Rating volatility

μjs: Ratings of oppenents within rating period

μjs: Rating deviations within rating period

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

τ=0.5: smaller values constrain the change in volatility over time, reasonable range between 0.3 and 1.2

ϵ=10^-6: Convergence bound

Returns tuple (μ, ϕ, σ): glicko2 rating, rating deviation and rating volatility for games in one rating period.

See: http://www.glicko.net/glicko/glicko2.pdf and Glickman, Mark E., "Dynamic paired comparison models with stochastic variances" (2001), Journal of Applied Statistics, 28, 673-689. (http://www.glicko.net/research/dpcmsv.pdf)

SkillRating.glicko_RD — Method

glicko_RD(r, RD, rjs::Vector, RDjs::Vector, sjs::Vector)
glicko_RD(r, RD, rj::T, RDj::T, sj::T) where T <: Real

Arguments:

r: Rating of player

RD: Rating deviation of player

rjs: Ratings of oppenents within rating period.

RDjs: Rating devations of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate a new glicko rating deviation for games within a rating period or for a single game.

See: http://www.glicko.net/glicko/glicko.pdf and Glickman, Mark E., "Parameter estimation in large dynamic paired comparison experiments" (1999) Applied Statistics, 48, 377-394 (http://www.glicko.net/research/glicko.pdf)

SkillRating.glicko_RD_increase — Method

glicko_RD_increase(rd; c = 63.2, maxrd=350)

Arguments:

RD: Rating deviation at end of rating period.

c: Increase in uncertainty between rating periods.

maxRD: Rating deviation of a player who does not play.

Calculate rating deviation after the end of a rating period.

See: http://www.glicko.net/glicko/glicko.pdf and Glickman, Mark E., "Parameter estimation in large dynamic paired comparison experiments" (1999) Applied Statistics, 48, 377-394 (http://www.glicko.net/research/glicko.pdf)

SkillRating.glicko_c — Method

glicko_c(rating_perods_to_max, maxrd, medianrd)

Calculate c, the increase in uncertainty between rating periods.

See: http://www.glicko.net/glicko/glicko.pdf and Glickman, Mark E., "Parameter estimation in large dynamic paired comparison experiments" (1999) Applied Statistics, 48, 377-394 (http://www.glicko.net/research/glicko.pdf)

SkillRating.glicko_expected — Method

elo_expected(RA, RB)

Arguments:

RA: Rating of player A

RB: Rating of player B

Calculate the expected outcome using glicko.

See: http://www.glicko.net/glicko/glicko.pdf and Glickman, Mark E., "Parameter estimation in large dynamic paired comparison experiments" (1999) Applied Statistics, 48, 377-394 (http://www.glicko.net/research/glicko.pdf)

SkillRating.glicko_rating — Method

glicko_rating(r, RD, rjs::Vector, RDjs::Vector, sjs::Vector)
glicko_rating(r, RD, rj::T, RDj::T, sj::T) where T <: Real

Arguments:

r: Rating of player

RD: Rating deviation of player

rjs: Ratings of oppenents within rating period.

RDjs: Rating devations of oppenents within rating period.

sjs: Game outcomes from the perspective of the "player": 1 - won, 0.5 - draw, 0 - lost.

Calculate a new glicko rating for games within a rating period or for a single game.

See: http://www.glicko.net/glicko/glicko.pdf and Glickman, Mark E., "Parameter estimation in large dynamic paired comparison experiments" (1999) Applied Statistics, 48, 377-394 (http://www.glicko.net/research/glicko.pdf)

SkillRating.glicko1_to_glicko2 — Method

glicko1_to_glicko2(r::real, RD::Real)

Convert ratings r and rating deviations RD from glicko1 (Elo-like) scale to glicko 2 scale (μ and ϕ).